+++ src/components/containers/og-card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import { useOGP } from '../util/hooks'
-import { appStore, PREFERENCE_DISPLAY_OGCARD } from '../stores'
+import { useOGP } from '../../util/hooks'
+import { appStore, PREFERENCE_DISPLAY_OGCARD } from '../../stores'
 import { OGCard } from '../presenters'
 
 export default ({
