+++ src/components/presenters/layout/header-logo/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Link } from 'react-router-dom'
-import logo from '../../../static/logo.png'
+import logo from '../../../../static/logo.png'
 import * as styles from './header-logo.css'
 
 const LogoImg = <img src={logo} width="64" height="64" alt="Mozuku" />
