+++ src/components/presenters/local-timeline-content/timeline/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import { Post as PostModel, AlbumFile } from '../../../models'
-import { Post } from '../../'
+import { Post as PostModel, AlbumFile } from '../../../../models'
+import { Post } from '../..'
 
 import * as styles from './timeline.css'
