+++ src/components/presenters/post/dummy-avatar/index.css
