+++ src/components/presenters/post/image/image.css.d.ts
